requirementes.txt
django 
psycopg2-binary
python-decouple
Pillow
djangorestframework